only update a summary file once per report step
this fixes the problem that well totals are off by a factor of
two. Note that the better solution would be to continue to write once
per substep but skipping report step writes if the adaptive time
stepper is used. since I've got no idea how to find this out I propose
this patch.